requirements.txt
streamlit>=1.29
torch>=2.1
torchvision>=0.16
numpy>=1.24
pillow>=10.0
matplotlib>=3.7
seaborn>=0.13
scikit-learn>=1.3
opencv-python-headless>=4.8
